Rename `tedge_api::entity` module to `tedge_api::mqtt_topics`
The aim of this module is to abstract the MQTT topics used by thin-edge.
Entities, such as devices and services, have to be described by the
`entity_store` module.
