+++ crates/core/tedge_api/src/lib.rs
@@ -6,12 +6,12 @@ pub mod topic;
 pub mod alarm;
 pub mod builder;
 pub mod data;
-pub mod entity;
 pub mod entity_store;
 pub mod event;
 pub mod group;
 pub mod health;
 pub mod measurement;
+pub mod mqtt_topics;
 pub mod parser;
 pub mod serialize;
 pub mod utils;
